+++ backend/src/main/java/com/example/backend/amenities/AmenitiesRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend.amenities;
 
-import com.example.backend.hotel.entity.Amenities;
+import com.example.backend.amenities.entity.Amenities;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AmenitiesRepository extends JpaRepository<Amenities, Long> {
