--- backend/src/main/java/com/example/backend/domain/Room.java
@@ -1,42 +1,0 @@
-//package com.example.backend.domain;
-//
-//import jakarta.persistence.*;
-//import lombok.Getter;
-//import lombok.NoArgsConstructor;
-//import lombok.Setter;
-//import java.math.BigDecimal;
-//import java.util.ArrayList;
-//import java.util.List;
-//
-//@Getter
-//@Setter
-//@Entity
-//@Table(name = "room")
-//@NoArgsConstructor
-//public class Room {
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//
-//    @Column(name = "price", nullable = false, precision = 10, scale = 2)
-//    private BigDecimal price;
-//
-//    @Column(name = "name", nullable = false, length = 20)
-//    private String name;
-//
-//    @Column(name = "view", length = 20)
-//    private String view;
-//
-//    @Column(name = "bed", length = 50)
-//    private String bed;
-//
-//    @Column(name = "rooms_guest", nullable = false, length = 50)
-//    private String roomsGuest;
-//
-//    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(name = "hotel_id", nullable = false)
-//    private Hotel hotel;
-//
-//    @OneToMany(mappedBy = "room", cascade = CascadeType.ALL)
-//    private List<Reservation> reservations = new ArrayList<>();
-//}
