--- backend/src/main/java/com/example/backend/domain/TravelPackage.java
@@ -1,44 +1,0 @@
-//package com.example.backend.domain;
-//
-//import jakarta.persistence.*;
-//import lombok.Getter;
-//import lombok.NoArgsConstructor;
-//import lombok.Setter;
-//import java.math.BigDecimal;
-//import java.time.LocalDate;
-//import java.util.ArrayList;
-//import java.util.List;
-//
-//@Getter @Setter
-//@Entity
-//@Table(name = "package")
-//@NoArgsConstructor
-//public class TravelPackage {
-//
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//
-//    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(name = "city_id", nullable = false)
-//    private City city;
-//
-//    @Column(nullable = false, length = 50)
-//    private String title;
-//
-//    @Column(nullable = false, length = 300)
-//    private String description;
-//
-//    @Column(nullable = false, precision = 10, scale = 2)
-//    private BigDecimal price;
-//
-//    @Column(name = "st_date", nullable = false)
-//    private LocalDate startDate;
-//
-//    @Column(name = "end_date", nullable = false)
-//    private LocalDate endDate;
-//
-//    @OneToMany(mappedBy = "travelPackage", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
-//    private List<PackageImage> images = new ArrayList<>();
-//
-//}
