--- backend/src/main/java/com/example/backend/repository/AmenitiesRepository.java
@@ -1,7 +1,0 @@
-//package com.example.backend.repository;
-//
-//import com.example.backend.hotel.entity.Amenities;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface AmenitiesRepository extends JpaRepository<Amenities, Long> {
-//}
