--- backend/src/main/java/com/example/backend/repository/CityRepository.java
@@ -1,7 +1,0 @@
-package com.example.backend.repository;
-
-import com.example.backend.common.City;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface  CityRepository  extends JpaRepository<City, Long> {
-}
