--- backend/src/main/java/com/example/backend/repository/FreebiesRepository.java
@@ -1,7 +1,0 @@
-package com.example.backend.repository;
-
-import com.example.backend.hotel.entity.Freebies;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FreebiesRepository extends JpaRepository<Freebies, Long> {
-}
