--- backend/src/main/java/com/example/backend/repository/HotelRepository.java
@@ -1,10 +1,0 @@
-//package com.example.backend.repository;
-//
-//import com.example.backend.hotel.entity.Hotel;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-//import org.springframework.stereotype.Repository;
-//
-//@Repository
-//public interface HotelRepository extends JpaRepository<Hotel, Long>, JpaSpecificationExecutor<Hotel> {
-//}
