--- backend/src/main/java/com/example/backend/repository/ReservationRepository.java
@@ -1,7 +1,0 @@
-package com.example.backend.repository;
-
-import com.example.backend.hotel.entity.Reservation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-}
