--- backend/src/main/java/com/example/backend/repository/RoomRepository.java
@@ -1,7 +1,0 @@
-package com.example.backend.repository;
-
-import com.example.backend.hotel.entity.Room;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomRepository extends JpaRepository<Room, Long> {
-}
