--- backend/src/main/java/com/example/backend/repository/TravelPackageRepository.java
@@ -1,7 +1,0 @@
-//package com.example.backend.repository;
-//
-//import com.example.backend.domain.TravelPackage;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface TravelPackageRepository extends JpaRepository<TravelPackage, Long> {
-//}
